moved mock code into package, mock camera test script is working again
