+++ mock_camera.py
@@ -1,448 +1,7 @@
 import pyqtgraph as pg
-import cv2 as cv
-import numpy as np
 import coorx
-from parallax.calibration import CameraTransform, StereoCameraTransform
-
-
-class PerspectiveTransform(coorx.BaseTransform):
-    """3D perspective or orthographic matrix transform using homogeneous coordinates.
-
-    Assumes a camera at the origin, looking toward the -Z axis.
-    The camera's top points toward +Y, and right points toward +X.
-
-    Points inside the perspective frustum are mapped to the range [-1, +1] along all three axes.
-    """
-    def __init__(self):
-        super().__init__(dims=(3, 3))
-        self.affine = coorx.AffineTransform(dims=(4, 4))
-
-    def _map(self, arr):
-        arr4 = np.empty((arr.shape[0], 4), dtype=arr.dtype)
-        arr4[:, :3] = arr
-        arr4[:, 3] = 1
-        out = self.affine._map(arr4)
-        return out[:, :3] / out[:, 3:4]
-
-    def set_ortho(self, left, right, bottom, top, znear, zfar):
-        """Set orthographic transform.
-        """
-        assert(right != left)
-        assert(bottom != top)
-        assert(znear != zfar)
-
-        M = np.zeros((4, 4), dtype=np.float32)
-        M[0, 0] = +2.0 / (right - left)
-        M[3, 0] = -(right + left) / float(right - left)
-        M[1, 1] = +2.0 / (top - bottom)
-        M[3, 1] = -(top + bottom) / float(top - bottom)
-        M[2, 2] = -2.0 / (zfar - znear)
-        M[3, 2] = -(zfar + znear) / float(zfar - znear)
-        M[3, 3] = 1.0
-        self.affine.matrix = M.T
-
-    def set_perspective(self, fovy, aspect, znear, zfar):
-        """Set the perspective
-
-        Parameters
-        ----------
-        fov : float
-            Field of view.
-        aspect : float
-            Aspect ratio.
-        near : float
-            Near location.
-        far : float
-            Far location.
-        """
-        assert(znear != zfar)
-        h = np.tan(fovy * np.pi / 360.0) * znear
-        w = h * aspect
-        self.set_frustum(-w, w, -h, h, znear, zfar)
-
-    def set_frustum(self, left, right, bottom, top, near, far):  # noqa
-        """Set the frustum
-        """
-        M = frustum(left, right, bottom, top, near, far)
-        self.affine.matrix = M.T
-
-
-def frustum(left, right, bottom, top, znear, zfar):
-    """Create view frustum
-
-    Parameters
-    ----------
-    left : float
-        Left coordinate of the field of view.
-    right : float
-        Right coordinate of the field of view.
-    bottom : float
-        Bottom coordinate of the field of view.
-    top : float
-        Top coordinate of the field of view.
-    znear : float
-        Near coordinate of the field of view.
-    zfar : float
-        Far coordinate of the field of view.
-
-    Returns
-    -------
-    M : ndarray
-        View frustum matrix (4x4).
-    """
-    assert(right != left)
-    assert(bottom != top)
-    assert(znear != zfar)
-
-    M = np.zeros((4, 4))
-    M[0, 0] = +2.0 * znear / float(right - left)
-    M[2, 0] = (right + left) / float(right - left)
-    M[1, 1] = +2.0 * znear / float(top - bottom)
-    M[2, 1] = (top + bottom) / float(top - bottom)
-    M[2, 2] = -(zfar + znear) / float(zfar - znear)
-    M[3, 2] = -2.0 * znear * zfar / float(zfar - znear)
-    M[2, 3] = -1.0
-    return M
-
-
-class LensDistortionTransform(coorx.BaseTransform):
-    """https://docs.opencv.org/4.x/dc/dbb/tutorial_py_calibration.html
-    """
-    def __init__(self, coeff=(0, 0, 0, 0, 0)):
-        super().__init__(dims=(2, 2))
-        self.coeff = coeff
-
-    def set_coeff(self, coeff):
-        self.coeff = coeff
-        self._update()
-
-    def _map(self, arr):
-        k1, k2, p1, p2, k3 = self.coeff
-
-        # radial distortion
-        r = np.linalg.norm(arr, axis=1)
-        dist = (1 + k1 * r**2 + k2 * r**4 + k3 * r**6)
-        out = arr * dist[:, None]
-
-        # tangential distortion
-        x = out[:, 0]
-        y = out[:, 1]
-        xy = x * y
-        r2 = r**2
-        out[:, 0] += 2 * p1 * xy + p2 * (r2 + 2 * x**2)
-        out[:, 1] += 2 * p2 * xy + p1 * (r2 + 2 * y**2)
-
-        return out
-
-
-class AxisSelectionEmbeddedTransform(coorx.BaseTransform):
-    def __init__(self, axes, transform, dims):
-        super().__init__(dims=dims)
-        self.axes = axes
-        self.subtr = transform
-
-    def _map(self, arr):
-        out = arr.copy()
-        out[:, self.axes] = self.subtr.map(arr[:, self.axes])
-        return out
-
-    def _imap(self, arr):
-        out = arr.copy()
-        out[:, axes] = self.subtr.imap(arr[:, axes])
-        return out
-
-
-class CameraTransform(coorx.CompositeTransform):
-    def __init__(self):
-        self.view = coorx.SRT3DTransform()
-        self.proj = PerspectiveTransform()
-        self.dist = LensDistortionTransform()
-        self.dist_embed = AxisSelectionEmbeddedTransform(axes=[0, 1], transform=self.dist, dims=(3, 3))
-        self.screen = coorx.STTransform(dims=(3, 3))
-        super().__init__([self.view, self.proj, self.dist_embed, self.screen])
-
-    def set_camera(self, center, look, fov, screen_size, up=(0, 0, 1), distortion=(0, 0, 0, 0, 0)):
-        center = np.asarray(center)
-        look = np.asarray(look)
-        up = np.asarray(up)
-
-        aspect_ratio = screen_size[0] / screen_size[1]
-        look_dist = np.linalg.norm(look - center)
-        forward = look - center
-        forward /= np.linalg.norm(forward)
-        right = np.cross(forward, up)
-        right /= np.linalg.norm(right)
-        up = np.cross(right, forward)
-        up /= np.linalg.norm(up)
-
-        pts1 = np.array([center, center + forward, center + right, center + up])
-        pts2 = np.array([[0, 0, 0], [0, 0, -1], [1, 0, 0], [0, 1, 0]])
-        self._view_mapping_err = self.view.set_mapping(pts1, pts2)
-
-
-        near = look_dist / 100
-        far = look_dist * 100
-
-        # set perspective with aspect=1 so that distortion is performed on
-        # isotropic coordinates
-        self.proj.set_perspective(fov, 1.0, near, far)
-
-        # correct for aspect in the screen transform instead
-        self.screen.set_mapping(
-            [[-1, -1 / aspect_ratio, -1], [1, 1 / aspect_ratio, 1]],
-            [[0, screen_size[1], 0], [screen_size[0], 0, 1]]
-        )
-
-        self.dist.set_coeff(distortion)
-
-
-class GraphicsItem:
-    def __init__(self, views):
-        self._transform = coorx.SRT3DTransform()
-        self.items = []
-        self.views = [GraphicsItemView(self, view) for view in views]
-
-    @property
-    def transform(self):
-        return self._transform
-
-    @transform.setter
-    def transform(self, tr):
-        self._transform = tr
-        for view in self.views:
-            view.update_transform()
-
-    def add_items(self, items):
-        self.items.extend(items)
-        self.render()
-
-    def render(self):
-        for view in self.views:
-            view.render()
-
-
-class GraphicsItemView:
-    def __init__(self, item, view):
-        self.item = item
-        self.view = view
-        self.full_transform = coorx.CompositeTransform([])
-        self.update_transform()
-        self.scene = view.scene()
-        self.full_transform.add_change_callback(self.transform_changed)
-
-    def update_transform(self):
-        self.full_transform.transforms = [self.item.transform, self.view.camera_tr]
-
-    def transform_changed(self, event):
-        self.render()
-
-    def render(self):
-        self.clear_graphics_items()
-        for item in self.item.items:
-            pts = None
-            if 'points' in item:
-                pts = self.full_transform.map(item['points'])
-                pts_xy = pts[:, :2]
-                pts_z = pts[:, 2]
-
-            if item['type'] == 'poly':
-                polygon = pg.QtGui.QPolygonF([pg.QtCore.QPointF(*pt) for pt in pts_xy])
-                gfx_item = pg.QtWidgets.QGraphicsPolygonItem(polygon)
-            elif item['type'] == 'line':
-                gfx_item = pg.QtWidgets.QGraphicsLineItem(*pts_xy.flatten())
-            elif item['type'] == 'plot':
-                gfx_item = pg.PlotCurveItem(x=pts_xy[:,0], y=pts_xy[:,1])
-            else:
-                raise TypeError(item['type'])
-
-            if 'pen' in item:
-                pen = pg.mkPen(item['pen'])
-                gfx_item.setPen(pen)
-
-            if 'brush' in item:
-                brush = pg.mkBrush(item['brush'])
-                gfx_item.setBrush(brush)
-
-            gfx_item.setZValue(-pts_z.mean())
-            item.setdefault('graphicsItems', {})
-            item['graphicsItems'][self] = gfx_item
-            self.scene.addItem(gfx_item)
-
-    def clear_graphics_items(self):
-        for item in self.item.items:
-            gfxitems = item.get('graphicsItems', {})
-            gfxitem = gfxitems.pop(self, None)
-            if gfxitem is not None:
-                self.scene.removeItem(gfxitem)
-
-
-class CheckerBoard(GraphicsItem):
-    def __init__(self, views, size, colors):
-        super().__init__(views)
-
-        for i in range(size):
-            for j in range(size):
-                self.items.append({
-                    'type': 'poly',
-                    'points': [[i, j, 0], [i+1, j, 0], [i+1, j+1, 0], [i, j+1, 0], [i, j, 0]],
-                    'pen': None,
-                    'brush': colors[(i  + j) % 2],
-                })
-        self.render()
-
-
-class Axis(GraphicsItem):
-    def __init__(self, views):
-        super().__init__(views)
-
-        self.items = [
-            {'type': 'line', 'points': [[0, 0, 0], [1, 0, 0]], 'pen': 'r'},
-            {'type': 'line', 'points': [[0, 0, 0], [0, 1, 0]], 'pen': 'g'},
-            {'type': 'line', 'points': [[0, 0, 0], [0, 0, 1]], 'pen': 'b'},
-        ]
-        self.render()
-
-
-class Electrode(GraphicsItem):
-    def __init__(self, views):    
-        super().__init__(views)
-
-        self.items = [
-            {'type': 'poly', 'pen': None, 'brush': 0.2, 'points': [
-                [0, 0, 0], [1, 0, 1], [1, 0, 100], [-1, 0, 100], [-1, 0, 1], [0, 0, 0]
-            ]},
-        ]
-        self.render()
-
-
-class GraphicsView3D(pg.GraphicsView):
-    def __init__(self, **kwds):
-        self.camera_tr = CameraTransform()
-        self.press_event = None
-        self.camera_params = {'look': [0, 0, 0], 'pitch': 30, 'yaw': 0, 'distance': 10, 'fov': 45, 'distortion': (0, 0, 0, 0, 0)}
-        super().__init__(**kwds)
-        self.setRenderHint(pg.QtGui.QPainter.Antialiasing)
-        self.set_camera(look=[0, 0, 0], pitch=30, yaw=0, distance=10, fov=45)
-
-    def set_camera(self, **kwds):
-        for k,v in kwds.items():
-            assert k in self.camera_params
-            self.camera_params[k] = v
-        self.update_camera()
-
-    def update_camera(self):
-        p = self.camera_params
-        look = np.asarray(p['look'])
-        pitch = p['pitch'] * np.pi/180
-        hdist = p['distance'] * np.cos(pitch)
-        yaw = p['yaw'] * np.pi/180
-        cam_pos = look + np.array([
-            hdist * np.cos(yaw),
-            hdist * np.sin(yaw),
-            p['distance'] * np.sin(pitch)
-        ])
-        self.camera_tr.set_camera(center=cam_pos, look=look, fov=p['fov'], screen_size=[self.width(), self.height()], distortion=p['distortion'])
-
-    def mousePressEvent(self, ev):
-        self.press_event = ev
-        self.last_mouse_pos = ev.pos()
-        ev.accept()
-
-    def mouseMoveEvent(self, ev):
-        if self.press_event is None:
-            return
-        dif = ev.pos() - self.last_mouse_pos        
-        self.last_mouse_pos = ev.pos()
-
-        self.camera_params['pitch'] = np.clip(self.camera_params['pitch'] + dif.y(), -90, 90)
-        self.camera_params['yaw'] -= dif.x()
-        self.update_camera()
-
-    def mouseReleaseEvent(self, ev):
-        self.press_event = None
-
-    def wheelEvent(self, event):
-        self.camera_params['distance'] *= 1.01**event.angleDelta().y()
-        self.update_camera()
-
-    def resizeEvent(self, ev):
-        super().resizeEvent(ev)
-        self.update_camera()
-
-    def get_array(self):
-        return pg.imageToArray(pg.QtGui.QImage(self.grab()), copy=True)
-
-
-def find_checker_corners(img, board_shape, show=False):
-    """https://docs.opencv.org/4.x/dc/dbb/tutorial_py_calibration.html
-    """
-    if show:
-        view = pg.image(img)
-
-    criteria = (cv.TERM_CRITERIA_EPS + cv.TERM_CRITERIA_MAX_ITER, 30, 0.001)
-
-    # prepare object points, like (0,0,0), (1,0,0), (2,0,0) ....,(6,5,0)
-    objp = np.zeros((board_shape[0] * board_shape[1], 3), dtype='float32')
-    objp[:, :2] = np.mgrid[0:board_shape[0], 0:board_shape[1]].T.reshape(-1, 2)
-
-    gray = cv.cvtColor(img, cv.COLOR_BGR2GRAY)
-
-    # Find the chess board corners
-    ret, corners = cv.findChessboardCorners(gray, board_shape, None)
-    if not ret:
-        return None, None
-
-    # If found, add object points, image points (after refining them)
-    imgp = cv.cornerSubPix(gray, corners, (11,11), (-1,-1), criteria)
-
-    if show:
-        plt = pg.ScatterPlotItem(x=imgp[:, 0, 1], y=imgp[:, 0, 0])
-        view.view.addItem(plt)
-
-    return objp, imgp
-
-
-def generate_calibration_data(view, n_images, cb_size):
-    imgp = []
-    objp = []
-    cam = view.camera_params.copy()
-    for i in range(n_images*4):
-        pitch = np.random.uniform(45, 89)
-        yaw = np.random.uniform(0, 360)
-        distance = np.random.uniform(5, 15)
-        view.set_camera(pitch=pitch, yaw=yaw, distance=distance)
-        op,ip = find_checker_corners(view.get_array(), cb_size)
-        if op is None:
-            continue
-        objp.append(op)
-        imgp.append(ip)
-        if len(imgp) >= n_images:
-            break
-        app.processEvents()
-    view.set_camera(**cam)
-    return objp, imgp
-
-
-def calibrate_camera(view, cb_size, n_images=40):
-    objp, imgp = generate_calibration_data(view, n_images=n_images, cb_size=cb_size)
-    ret, mtx, dist, rvecs, tvecs = cv.calibrateCamera(objp, imgp, (view.width(), view.height()), None, None)
-    return ret, mtx, dist, rvecs, tvecs
-
-
-def undistort_image(img, mtx, dist, crop=False):
-    h, w = img.shape[:2]
-    new_camera_mtx, roi = cv.getOptimalNewCameraMatrix(mtx, dist, (w,h), 1, (w,h))
-
-    # undistort
-    mapx, mapy = cv.initUndistortRectifyMap(mtx, dist, None, new_camera_mtx, (w,h), 5)
-    dst = cv.remap(img, mapx, mapy, cv.INTER_LINEAR)
-
-    # crop the image
-    if crop:
-        x, y, w, h = roi
-        dst = dst[y:y+h, x:x+w]
-    return dst
-
+from parallax.mock_sim import GraphicsView3D, CheckerBoard, Axis, Electrode, calibrate_camera, undistort_image
+from parallax.lib import find_checker_corners
 
 class StereoView(pg.QtWidgets.QWidget):
     def __init__(self, parent=None, background=(128, 128, 128)):
@@ -470,7 +29,7 @@ if __name__ == '__main__':
     camera_params = dict(
         pitch=30,
         distance=15,
-        distortion=(0, 0, 0, 0, 0),
+        distortion=(-0.1, 0, 0, 0, 0),
     )
     win.set_camera(0, yaw=-5, **camera_params)
     win.set_camera(1, yaw=5, **camera_params)
@@ -481,20 +40,16 @@ if __name__ == '__main__':
     axis = Axis(views=win.views)
 
     s = 0.1
-    electrodes = []
-    for i in range(4):
-        e = Electrode(win.views)
-        electrodes.append(e)
-        # e.transform.set_scale([s, s, s])
-        # e.transform.set_offset([-5, 0, 5])
-        # e.transform.rotate(45, [1, 0, 0])
-        # e.transform.rotate(15 * i, [0, 0, 1])
+    # electrodes = []
+    # for i in range(4):
+    #     e = Electrode(win.views)
+    #     electrodes.append(e)
 
-        e.transform = coorx.AffineTransform(dims=(3, 3))
-        e.transform.scale([s, s, s])
-        e.transform.rotate(60, [1, 0, 0])
-        e.transform.translate([0, 1, 1])
-        e.transform.rotate(15*i, [0, 0, 1])
+    #     e.transform = coorx.AffineTransform(dims=(3, 3))
+    #     e.transform.scale([s, s, s])
+    #     e.transform.rotate(60, [1, 0, 0])
+    #     e.transform.translate([0, 1, 1])
+    #     e.transform.rotate(15*i, [0, 0, 1])
 
 
     # tr = coorx.AffineTransform(dims=(3, 3))
@@ -522,7 +77,7 @@ if __name__ == '__main__':
         v1 = pg.image(img)
         v2 = pg.image(uimg)
         tr = coorx.AffineTransform(matrix=mtx[:2, :2], offset=mtx[:2, 2])
-        ltr = LensDistortionTransform(dist[0])
+        ltr = coorx.nonlinear.LensDistortionTransform(dist[0])
         ttr = coorx.CompositeTransform([tr.inverse, ltr, tr])
 
         objp, imgp = find_checker_corners(uimg, board_shape=(cb_size-1, cb_size-1))
@@ -530,8 +85,8 @@ if __name__ == '__main__':
 
         distorted_pts = ttr.map(undistorted_pts)
 
-        s1 = pg.ScatterPlotItem(x=undistorted_pts[:,1], y=undistorted_pts[:,0])
+        s1 = pg.ScatterPlotItem(x=undistorted_pts[:,1], y=undistorted_pts[:,0], brush='r', pen=None)
         v2.view.addItem(s1)
 
-        s2 = pg.ScatterPlotItem(x=distorted_pts[:,1], y=distorted_pts[:,0])
+        s2 = pg.ScatterPlotItem(x=distorted_pts[:,1], y=distorted_pts[:,0], brush='r', pen=None)
         v1.view.addItem(s2)
