+++ parallax/calibration.py
@@ -2,6 +2,7 @@
 
 import numpy as np
 import cv2 as cv
+import coorx
 from . import lib
 from .helper import WF, HF
